requirements.txt
# Python 3.9.25
# PyTorch with CUDA 11.8 — install separately:
#   conda install pytorch==2.5.1 torchvision==0.20.1 pytorch-cuda=11.8 -c pytorch -c nvidia

torch==2.5.1
torchvision==0.20.1
numpy==2.0.1
Pillow==11.1.0
matplotlib==3.9.4
scipy==1.13.1
tqdm==4.67.3